requirements.txt
Flask==0.11.1
httplib2==0.9.2
Jinja2==2.8
matplotlib==1.5.1
numpy==1.11.1
pandas==0.18.1
peewee==2.9.1
prettytable==0.7.2
google_api_python_client==1.6.2
python_dateutil==2.6.0
scipy==0.19.0
scikit_learn==0.18.1
PyYAML==3.12
